+++ index.ts
@@ -17,6 +17,7 @@ import feedbackRouter from "./routers/feedbackRouter";
 import paymentRouter from "./routers/paymentRouter";
 import orederRouter from "./routers/orderRouter";
 import productsRouter from "./routers/productsRouter";
+import searchRouter from "./routers/searchRouter";
 
 dotenv.config();
 
@@ -171,41 +172,7 @@ app.use("/api/v1/order", orederRouter);
 
 app.use("/api/v1/products", productsRouter);
 
-app.get("/api/v1/search", async (req: Request, res: Response) => {
-  const keyword = req.query.kw as string;
-
-  if (!keyword || keyword.trim() === "") {
-    res.status(400).send({ error: "Invalid keyword" });
-    return;
-  }
-
-  try {
-    const lowerKeyword = keyword.toLowerCase();
-    const products = await Product.findAll({
-      where: Sequelize.where(
-        Sequelize.fn("lower", Sequelize.col("product_name")),
-        "LIKE",
-        `%${lowerKeyword}%`
-      ),
-      attributes: [
-        "product_name",
-        "product_price",
-        "product_stock",
-        "image_url",
-      ],
-    });
-
-    if (!products) {
-      res.status(404).send({ error: "No products found" });
-      return;
-    }
-
-    res.status(200).json(products);
-  } catch (error) {
-    console.error(error);
-    res.status(500).send({ error: "Error retrieving product data" });
-  }
-});
+app.use("/api/v1/search", searchRouter);
 
 app.get("/api/v1/popular", async (req: Request, res: Response) => {
   const n = 5; // 상위 n개의 상품
